requirements.txt
Flask==2.2.2
imgaug==0.4.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
matplotlib==3.4.3
matplotlib-inline==0.1.2
numpy==1.20.3
opencv-python==4.5.3.56
Pillow==8.2.0
pygame==2.1.0
PyYAML==5.4.1
scikit-image==0.16.2
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
sklearn==0.0
tensorboardX==2.4
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
virtualenv==20.0.17
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
Werkzeug==2.2.2